# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/personal-information": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "personal-information"
                ],
                "summary": "Insert Personal Information",
                "parameters": [
                    {
                        "description": "Body Request",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.InsertPersonalInformationReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Personal Information Details",
                        "schema": {
                            "$ref": "#/definitions/dto.BaseResp"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorSchema"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorSchema"
                        }
                    }
                }
            }
        },
        "/api/personal-information/{id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "personal-information"
                ],
                "summary": "Update Personal Information",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Personal Information ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Body Request",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdatePersonalInformationReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Personal Information Details",
                        "schema": {
                            "$ref": "#/definitions/dto.BaseResp"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorSchema"
                        }
                    }
                }
            }
        },
        "/api/skill": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "skill"
                ],
                "summary": "Insert Skill",
                "parameters": [
                    {
                        "description": "Body Request",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.InsertSkillReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Skill Details",
                        "schema": {
                            "$ref": "#/definitions/dto.BaseResp"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorSchema"
                        }
                    }
                }
            }
        },
        "/api/skill/{id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "skill"
                ],
                "summary": "Update Skill",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Skill ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Body Request",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateSkillReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Skill resp",
                        "schema": {
                            "$ref": "#/definitions/dto.BaseResp"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorSchema"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.BaseResp": {
            "type": "object",
            "properties": {
                "error_schema": {
                    "$ref": "#/definitions/dto.ErrorSchema"
                },
                "output_schema": {
                    "description": "OutputSchema fleksibel (bisa berisi apa saja)"
                }
            }
        },
...
        "dto.ErrorMessage": {
            "type": "object",
            "properties": {
                "en": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                }
            }
        },
        "dto.ErrorSchema": {
            "type": "object",
            "properties": {
                "error_code": {
                    "type": "string"
                },
                "error_message": {
                    "$ref": "#/definitions/dto.ErrorMessage"
                }
            }
        },
...
        "dto.InsertPersonalInformationReq": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "date_of_birth": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "first_name": {
                    "type": "string"
                },
                "last_name": {
                    "type": "string"
                },
                "phone_number": {
                    "type": "string"
                },
                "summary": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "dto.InsertSkillReq": {
            "type": "object",
            "properties": {
                "level": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
...
        "dto.UpdatePersonalInformationReq": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "date_of_birth": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "first_name": {
                    "type": "string"
                },
                "last_name": {
                    "type": "string"
                },
                "phone_number": {
                    "type": "string"
                },
                "summary": {
                    "type": "string"
                }
            }
        },
        "dto.UpdateSkillReq": {
            "type": "object",
            "properties": {
                "level": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}